monkey/pyproject.toml
[project]
name = "monkey"
version = "0.1.0"
description = "A web crawler agent using Selenium and Google ADK"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "selenium>=4.30.0,<5.0.0",
    "webdriver-manager>=4.0.2,<5.0.0",
    "pillow>=11.1.0,<12.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "google-adk>=0.0.2",
    "google-genai>=1.5.0,<2.0.0",
    "google-cloud-bigquery>=3.31.0,<4.0.0",
    "google-cloud-aiplatform[agent-engines] @ git+https://github.com/googleapis/python-aiplatform.git@copybara_738852226",
    "absl-py>=2.2.2,<3.0.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api" 